support decimals

+++ contracts/interface/IMB.sol
@@ -3,6 +3,7 @@
 pragma solidity ^0.8.0;
 
 interface IMB {
-    function mint(address to, uint amount) external returns(bool);
-    function burn(address from, uint amount) external returns(bool);
+    function mint(address to, uint256 amount) external returns (bool);
+
+    function burn(address from, uint256 amount) external returns (bool);
 }
